requirements.txt
playwright
playwright-stealth
requests
beautifulsoup4
python-dotenv
